feat: DSR support
