+++ plugin/package.json
@@ -2,7 +2,7 @@
   "name": "@netlify/plugin-gatsby",
   "version": "1.0.3",
   "description": "Netlify Build plugin - Run Gatsby seamlessly on Netlify",
-  "source": "src/index.js",
+  "source": "src/index.ts",
   "main": "lib/main.js",
   "module": "lib/module.js",
   "files": [
@@ -46,14 +46,16 @@
     "cookie": "^0.4.1",
     "download": "^8.0.0",
     "fs-extra": "^10.0.0",
+    "linkfs": "^2.1.0",
     "multer": "^1.4.2",
     "node-fetch": "^2.6.1",
     "path-to-regexp": "^0.1.7",
     "tempy": "^1.0.0"
   },
   "devDependencies": {
+    "@types/fs-extra": "^9.0.12",
     "@types/multer": "^1.4.7",
-    "gatsby": "^3.14.0-next.0",
+    "gatsby": "^3.11.0-alpha-qe-sm.70",
     "npm-run-all": "^4.1.5",
     "parcel": "^2.0.0-rc.0",
     "release-it": "^14.8.0",
