+++ plugin/src/templates/api/gatsbyFunction.ts
@@ -3,13 +3,18 @@ import bodyParser from 'co-body'
 import multer from 'multer'
 import path from 'path'
 import { existsSync } from 'fs'
-import { proxyRequest } from './functions'
-import { AugmentedGatsbyFunctionResponse } from './createResponseObject'
-import { AugmentedGatsbyFunctionRequest } from './createRequestObject'
+import { proxyRequest } from './utils'
+import {
+  AugmentedGatsbyFunctionResponse,
+  AugmentedGatsbyFunctionRequest,
+} from './utils'
 import { HandlerEvent } from '@netlify/functions'
 
 const parseForm = multer().any()
 
+/**
+ * Execute a Gatsby function
+ */
 export async function gatsbyFunction(
   req: AugmentedGatsbyFunctionRequest,
   res: AugmentedGatsbyFunctionResponse<any>,
@@ -17,7 +22,7 @@ export async function gatsbyFunction(
 ) {
   // Multipart form data middleware. because co-body can't handle it
 
-  // @ts-ignore As we're using a fake Express handler we need to ignore the type
+  // @ts-ignore As we're using a fake Express handler we need to ignore the type to keep multer happy
   await new Promise((next) => parseForm(req, res, next))
   try {
     // If req.body is populated then it was multipart data
@@ -33,8 +38,7 @@ export async function gatsbyFunction(
     console.log('Error parsing body', e, req)
   }
 
-  //  Strip "/api/" from path
-  const pathFragment = decodeURIComponent(req.url.substr(5))
+  let pathFragment = decodeURIComponent(req.url)
 
   let functions
   try {
@@ -47,18 +51,69 @@ export async function gatsbyFunction(
     }
   }
 
+  const ssrRoutes = new Map()
+
+  functions.forEach((fn) => {
+    // Functions that start with "_ssr" are actually SSR page routes
+    if (
+      fn.functionRoute.startsWith(`_ssr/`) &&
+      !fn.functionRoute.startsWith(`_ssr/page-data/`)
+    ) {
+      let pathName: string | RegExp = fn.functionRoute
+
+      let pageDataRoute: string | RegExp = fn.functionRoute.replace(
+        `_ssr/`,
+        `_ssr/page-data/`,
+      )
+      const keys = []
+      if (fn.matchPath) {
+        pathName = pathToRegexp(fn.matchPath, keys, {})
+        pageDataRoute = pathToRegexp(
+          fn.matchPath.replace(`_ssr/`, `_ssr/page-data/`) + `/page-data.json`,
+          [],
+          {},
+        )
+      }
+
+      ssrRoutes.set(pathName, {
+        apiRoute: `/api/${fn.functionRoute}`,
+        params: keys.map((key) => key.name),
+      })
+      ssrRoutes.set(pageDataRoute, {
+        apiRoute: `/api/${fn.functionRoute.replace(`_ssr`, `_ssr/page-data`)}`,
+        params: keys.map((key) => key.name),
+      })
+    }
+  })
+
   // Find the matching function, given a path. Based on Gatsby Functions dev server implementation
   // https://github.com/gatsbyjs/gatsby/blob/master/packages/gatsby/src/internal-plugins/functions/gatsby-node.ts
 
+  for (const [ssrPath, ssrFn] of ssrRoutes) {
+    const match = `_ssr${pathFragment}`.match(ssrPath)
+    if (match) {
+      let url = ssrFn.apiRoute
+      ssrFn.params.forEach((param, index) => {
+        url = url.replace(`[${param}]`, match[index + 1])
+      })
+
+      pathFragment = url
+      res.setHeader('content-type', 'text/html')
+      break
+    }
+  }
+
+  pathFragment = pathFragment.replace('/api/', '')
+
   // Check first for exact matches.
   let functionObj = functions.find(
-    ({ apiRoute, functionRoute }) =>
-      (functionRoute || apiRoute) === pathFragment,
+    ({ functionRoute }) => functionRoute === pathFragment,
   )
 
   if (!functionObj) {
     // Check if there's any matchPaths that match.
     // We loop until we find the first match.
+
     functions.some((f) => {
       let exp
       const keys = []
@@ -84,19 +139,22 @@ export async function gatsbyFunction(
     const start = Date.now()
 
     const pathToFunction = process.env.NETLIFY_DEV
-      ? functionObj.absoluteCompiledFilePath
-      : path.join(
+      ? // During develop, the absolute path is correct
+        functionObj.absoluteCompiledFilePath
+      : // ...otherwise we need to use a relative path, as we're in a lambda
+        path.join(
           __dirname,
           '..',
           '..',
           '..',
+          // ...We got there in the end
           '.cache',
           'functions',
           functionObj.relativeCompiledFilePath,
         )
 
     if (process.env.NETLIFY_DEV && !existsSync(pathToFunction)) {
-      // Functions are lazily-compiled, so we proxy the first request to the gatsby develop server
+      // Functions are sometimes lazily-compiled, so we check and proxy the request if needed
       console.log(
         'No compiled function found. Proxying to gatsby develop server',
       )
@@ -104,6 +162,7 @@ export async function gatsbyFunction(
     }
 
     try {
+      // Make sure it's hot and fresh from the filesystem
       delete require.cache[require.resolve(pathToFunction)]
       const fn = require(pathToFunction)
